add TE context to sdxl
